# api/docs/openapi.yaml
openapi: 3.0.0
info:
  title: User API Service
  description: API для социальной сети с пользователями, друзьями и постами
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
...
paths:
...
  /friend/add:
    post:
      tags:
        - Friends
      summary: Добавить друга
      description: Добавляет пользователя в друзья
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FriendRequest'
            example:
              friend_id: 2
      responses:
        '200':
          description: Друг успешно добавлен
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Friend added successfully"
        '400':
          description: Неверные данные или пользователи уже друзья
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Пользователь не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /friend/delete:
    post:
      tags:
        - Friends
      summary: Удалить друга
      description: Удаляет пользователя из друзей
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FriendRequest'
            example:
              friend_id: 2
      responses:
        '200':
          description: Друг успешно удален
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Friend deleted successfully"
        '400':
          description: Неверные данные или пользователи не являются друзьями
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Дружба не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /friends:
    get:
      tags:
        - Friends
      summary: Получить список друзей
      description: Возвращает список друзей пользователя
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Список друзей
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FriendResponse'
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /friend/status:
    get:
      tags:
        - Friends
      summary: Получить статус дружбы
      description: Возвращает статус дружбы с указанным пользователем
      security:
        - BearerAuth: []
      parameters:
        - name: friend_id
          in: query
          required: true
          description: ID пользователя
          schema:
            type: integer
            format: int64
            example: 2
      responses:
        '200':
          description: Статус дружбы
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FriendshipStatus'
        '400':
          description: Неверный ID пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /post/create:
    post:
      tags:
        - Posts
      summary: Создать пост
      description: Создает новый пост пользователя
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePostRequest'
            example:
              title: "Мой первый пост"
              content: "Это содержимое моего первого поста в социальной сети"
      responses:
        '201':
          description: Пост успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '400':
          description: Неверные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /post/get/{id}:
    get:
      tags:
        - Posts
      summary: Получить пост
      description: Возвращает пост по ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID поста
          schema:
            type: integer
            format: int64
            example: 1
      responses:
        '200':
          description: Информация о посте
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostResponse'
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Пост не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /post/update/{id}:
    put:
      tags:
        - Posts
      summary: Обновить пост
      description: Обновляет существующий пост
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID поста
          schema:
            type: integer
            format: int64
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePostRequest'
            example:
              title: "Обновленный заголовок"
              content: "Обновленное содержимое поста"
      responses:
        '200':
          description: Пост успешно обновлен
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Post updated successfully"
        '400':
          description: Неверные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Пост не найден или доступ запрещен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /post/delete/{id}:
    delete:
      tags:
        - Posts
      summary: Удалить пост
      description: Удаляет пост пользователя
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID поста
          schema:
            type: integer
            format: int64
            example: 1
      responses:
        '200':
          description: Пост успешно удален
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Post deleted successfully"
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Пост не найден или доступ запрещен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /posts:
    get:
      tags:
        - Posts
      summary: Получить посты пользователя
      description: Возвращает посты указанного пользователя
      security:
        - BearerAuth: []
      parameters:
        - name: user_id
          in: query
          required: false
          description: ID пользователя (по умолчанию - текущий)
          schema:
            type: integer
            format: int64
            example: 1
        - name: page
          in: query
          required: false
          description: Номер страницы
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: page_size
          in: query
          required: false
          description: Размер страницы
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Список постов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedResponse'
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /post/feed:
    get:
      tags:
        - Posts
      summary: Лента постов друзей
      description: Возвращает ленту постов друзей пользователя
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          description: Номер страницы
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: page_size
          in: query
          required: false
          description: Размер страницы
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Лента постов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedResponse'
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ID пользователя
          example: 1
        username:
          type: string
          description: Имя пользователя
          example: "john_doe"
        email:
          type: string
          format: email
          description: Email пользователя
          example: "john@example.com"
        first_name:
          type: string
          description: Имя пользователя
          example: "John"
        last_name:
          type: string
          description: Фамилия пользователя
          example: "Doe"
        birth_date:
          type: string
          format: date
          description: Дата рождения
          example: "1990-01-15"
        gender:
          type: string
          description: Пол пользователя
          enum:
            - male
            - female
            - unknown
          example: "male"
        interests:
          type: string
          description: Интересы пользователя
          example: "programming, reading, sports"
        city:
          type: string
          description: Город пользователя
          example: "New York"
        created_at:
          type: string
          format: date-time
          description: Дата и время создания пользователя
          example: "2023-12-19T10:30:00Z"
...
    FriendRequest:
      type: object
      required:
        - friend_id
      properties:
        friend_id:
          type: integer
          format: int64
          description: ID пользователя для добавления/удаления в друзья
          example: 2
    FriendResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ID записи о дружбе
          example: 1
        user_id:
          type: integer
          format: int64
          description: ID пользователя
          example: 1
        friend_id:
          type: integer
          format: int64
          description: ID друга
          example: 2
        friend:
          $ref: '#/components/schemas/UserResponse'
        created_at:
          type: string
          format: date-time
          description: Дата и время добавления в друзья
          example: "2023-12-19T10:30:00Z"
    FriendshipStatus:
      type: object
      properties:
        is_friend:
          type: boolean
          description: Являются ли пользователи друзьями
          example: true
        is_pending:
          type: boolean
          description: Ожидает ли запрос на дружбу подтверждения
          example: false
    CreatePostRequest:
      type: object
      required:
        - title
        - content
      properties:
        title:
          type: string
          description: Заголовок поста
          maxLength: 255
          example: "Мой первый пост"
        content:
          type: string
          description: Содержимое поста
          example: "Это содержимое моего первого поста в социальной сети"
    Post:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ID поста
          example: 1
        user_id:
          type: integer
          format: int64
          description: ID пользователя
          example: 1
        title:
          type: string
          description: Заголовок поста
          example: "Мой первый пост"
        content:
          type: string
          description: Содержимое поста
          example: "Это содержимое моего первого поста в социальной сети"
        created_at:
          type: string
          format: date-time
          description: Дата и время создания поста
          example: "2023-12-19T10:30:00Z"
        updated_at:
          type: string
          format: date-time
          description: Дата и время обновления поста
          example: "2023-12-19T10:30:00Z"
    PostResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ID поста
          example: 1
        user_id:
          type: integer
          format: int64
          description: ID пользователя
          example: 1
        user:
          $ref: '#/components/schemas/UserResponse'
        title:
          type: string
          description: Заголовок поста
          example: "Мой первый пост"
        content:
          type: string
          description: Содержимое поста
          example: "Это содержимое моего первого поста в социальной сети"
        created_at:
          type: string
          format: date-time
          description: Дата и время создания поста
          example: "2023-12-19T10:30:00Z"
        updated_at:
          type: string
          format: date-time
          description: Дата и время обновления поста
          example: "2023-12-19T10:30:00Z"
    UpdatePostRequest:
      type: object
      properties:
        title:
          type: string
          description: Заголовок поста
          maxLength: 255
          example: "Обновленный заголовок"
          nullable: true
        content:
          type: string
          description: Содержимое поста
          example: "Обновленное содержимое поста"
          nullable: true
    FeedResponse:
      type: object
      properties:
        posts:
          type: array
          items:
            $ref: '#/components/schemas/PostResponse'
        total:
          type: integer
          description: Общее количество постов
          example: 150
        page:
          type: integer
          description: Текущая страница
          example: 1
        pages:
          type: integer
          description: Общее количество страниц
          example: 8
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Сообщение об ошибке
          example: "Invalid credentials"
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Введите JWT токен в формате "Bearer {token}"
...